🎨 Extract theme to constants folder
