+++ src/App.tsx
@@ -1,28 +1,14 @@
 import React from 'react';
-import { DefaultTheme, Provider as PaperProvider } from 'react-native-paper';
+
+// React native paper provider
+import { Provider as PaperProvider } from 'react-native-paper';
+
+// React navigation
+import 'react-native-gesture-handler';
 import { NavigationContainer } from '@react-navigation/native';
 
 import Routes from './routes';
-
-// Customizar tema posteriormente
-const theme = {
-  ...DefaultTheme,
-  dark: false,
-  roundness: 4,
-  colors: {
-    ...DefaultTheme.colors,
-    primary: '#6200ee',
-    accent: '#03dac4',
-    background: '#f6f6f6',
-    surface: 'white',
-    error: '#B00020',
-    text: 'black',
-    onSurface: '#000000',
-  },
-  animation: {
-    scale: 1.0,
-  },
-};
+import { theme } from './constants';
 
 export default function App(): JSX.Element {
   return (
