+++ src/constants/index.ts
@@ -1,0 +1,5 @@
+import theme from './theme';
+
+export {
+  theme,
+};
